refactor to new component; fix issue with adding matrix column after data exists
